Move ServiveAdvertiser interface to servicediscovery package
Advertising and discovery are two sides of a same coin.
