+++ babble/src/main/java/io/mosaicnetworks/babble/servicediscovery/mdns/MdnsAdvertiser.java
@@ -32,7 +32,7 @@ import java.io.IOException;
 import java.util.Objects;
 
 import io.mosaicnetworks.babble.node.GroupDescriptor;
-import io.mosaicnetworks.babble.service.ServiceAdvertiser;
+import io.mosaicnetworks.babble.servicediscovery.ServiceAdvertiser;
 import io.mosaicnetworks.babble.utils.RandomString;
 
 public class MdnsAdvertiser implements ServiceAdvertiser {
